+++ system/logger.go
@@ -1,0 +1,11 @@
+package system
+
+import "github.com/sirupsen/logrus"
+
+var logger *logrus.Entry
+
+func init() {
+	logger = logrus.WithFields(logrus.Fields{
+		"component": "system",
+	})
+}